% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{compute_datetime_info}
\alias{compute_datetime_info}
\title{Compute a named list with datetime ranges and step.}
\usage{
compute_datetime_info(datetime_from, datetime_to, datetime_step)
}
\arguments{
\item{datetime_from}{The initial datetime point of the range as string, in
format \code{YYYY-MM-DD HH:MM:SS}.}

\item{datetime_to}{The final datetime point of the range as string, in format
\code{YYYY-MM-DD HH:MM:SS}.}

\item{datetime_step}{The size of the step within the range, with 1 being each
hour, 2 every other hour, etc.}
}
\value{
A named list with elements \code{datetime_from}, \code{datetime_to}
and \code{datetime_step}.
}
\description{
Helper function to get formatted data for use in other
functions.
}
