% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{datetime_to_index}
\alias{datetime_to_index}
\title{Convert a datetime to an index in the WIND Toolkit dataset}
\usage{
datetime_to_index(datetime)
}
\arguments{
\item{datetime}{The datetime point as POSIX or string, in format
\code{YYYY-MM-DD HH:MM:SS}.}
}
\value{
The index on the WIND Toolkit dataset for the datetime.
}
\description{
Given a datetime, compute the corresponding index on the WIND
Toolkit dataset.
}
