% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{datetime_to_indices}
\alias{datetime_to_indices}
\title{Convert a datetime range to the indices range in the WIND Toolkit dataset}
\usage{
datetime_to_indices(datetime_from, datetime_to)
}
\arguments{
\item{datetime_from}{The initial datetime point of the range as POSIX or
string, in format \code{YYYY-MM-DD HH:MM:SS}.}

\item{datetime_to}{The final datetime point of the range as POSIX or string,
in format \code{YYYY-MM-DD HH:MM:SS}.}
}
\value{
A named list with elements \code{t_from} and \code{t_to} for the
corresponding indices range on the WIND Toolkit dataset.
}
\description{
Given two datetime (from and to), compute the corresponding
indices range on the WIND Toolkit dataset.
}
