% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_dataset}
\alias{get_dataset}
\title{Get a portion of the WIND Toolkit dataset as a tidy dataframe for a datetime
range and a geographic nearest point.}
\usage{
get_dataset(datasets, api_key, dataset_titles, datetime_info, latitude,
  longitude)
}
\arguments{
\item{datasets}{The dataframe with datasets retrieved from
\code{get_datasets()}}

\item{api_key}{API key to access the dataset.}

\item{dataset_titles}{A string vector containing the dataset titles to
retrieve, e.g.
\code{c("windspeed_80m", "temperature_80m", "relativehumidity_2m")}. Avoid
the special datasets \code{"datetime"} and \code{"coordinates"}, which are
added by default in the process.}

\item{datetime_info}{A named list (as formatted by
\code{compute_datetime_info()}) with the datetime range and step.}

\item{latitude}{The latitude of the point (WGS84).}

\item{longitude}{The longitude of the point (WGS84).}
}
\value{
A dataframe (tibble) in tidy format with the values of each
information requested for the datetime range and the geographic nearest
point.
}
\description{
Get a portion of the WIND Toolkit dataset as a tidy dataframe for a datetime
range and a geographic nearest point.
}
