% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_dataset_url}
\alias{get_dataset_url}
\title{Get the URL of a specific dataset available via the WIND Toolkit API.}
\usage{
get_dataset_url(datasets, dataset_title)
}
\arguments{
\item{datasets}{The dataframe with datasets retrieved from
\code{get_datasets()}}

\item{dataset_title}{A string with the specific title of the dataset, e.g.
\code{"windspeed_80m"}, \code{"temperature_80m"},
\code{"relativehumidity_2m"}.}
}
\value{
A string with the URL of the dataset. Empty if the
\code{dataset_title} is not valid.
}
\description{
Get the URL of a specific dataset available via the WIND Toolkit API.
}
