% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_endpoint}
\alias{is_endpoint}
\title{Check if the endpoint is a valid URL}
\usage{
is_endpoint(endpoint)
}
\arguments{
\item{endpoint}{The endpoint URL.}
}
\value{
TRUE if the endpoint starts with \code{"http://"} or
\code{"https://"}, FALSE otherwise.
}
\description{
Validates that the parameter \code{endpoint} starts with either
\code{"http://"} or \code{"https://"}.
}
