% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{is_nrel_host}
\alias{is_nrel_host}
\title{Check if the host is a valid NREL host}
\usage{
is_nrel_host(host)
}
\arguments{
\item{host}{The host address of the NREL HDF5 file.}
}
\value{
TRUE if the host is a valid NREL host, FALSE otherwise.
}
\description{
Validates that the parameter \code{host} starts with
\code{"/nrel/"} and ends with \code{".h5"}.
}
