% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{latlong_to_LCC}
\alias{latlong_to_LCC}
\title{Project latitude and longitude to Lambert Conformal Conic coordinates}
\usage{
latlong_to_LCC(latitude, longitude)
}
\arguments{
\item{latitude}{The latitude of the point (WGS84).}

\item{longitude}{The longitude of the point (WGS84).}
}
\value{
A named list with elements \code{coord_y} for the latitude and
\code{coord_x} for the longitude, both as LCC coordinates.
}
\description{
Given the latitude and longitude of a point, compute the
projection to Lambert Conformal Conic coordinates.
}
