% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{parse_response_content}
\alias{parse_response_content}
\title{Parse the content of an API response}
\usage{
parse_response_content(response)
}
\arguments{
\item{response}{Object with the API response.}
}
\value{
The content of the JSON API response.
}
\description{
Get the content of a JSON API response.
}
